requirements.txt
Flask==3.0.0
google-genai
google~=3.0.0

python-dotenv~=1.2.1
requests~=2.32.5
Flask-Migrate~=4.1.0
Flask-Login~=0.6.3
openai~=2.14.0
alembic~=1.17.2
SQLAlchemy~=2.0.45
Werkzeug~=3.0.1
pydantic~=2.12.5